chore: updating to work with the new @types/node

+++ packages/e2e-test/src/commands/run.ts
@@ -350,6 +350,7 @@ async function testAppServe(pluginName: string, appDir: string) {
       GITHUB_TOKEN: 'abc',
     },
   });
+
   Browser.localhost('localhost', 3000);
 
   let successful = false;
@@ -378,7 +379,9 @@ async function testAppServe(pluginName: string, appDir: string) {
     }
   } finally {
     // Kill entire process group, otherwise we'll end up with hanging serve processes
-    killTree(startApp.pid);
+    if (startApp.pid) {
+      killTree(startApp.pid);
+    }
   }
 
   try {
@@ -457,7 +460,9 @@ async function testBackendStart(appDir: string, isPostgres: boolean) {
   } finally {
     print('Stopping the child process');
     // Kill entire process group, otherwise we'll end up with hanging serve processes
-    killTree(child.pid);
+    if (child.pid) {
+      killTree(child.pid);
+    }
   }
 
   try {
